Version 1.3
Version 1.3:

Rearrangetabsbyfirstletter renamed to Rearrangetabsbyfirstletters: If called with no parameters, considers the first character of the file name, as before. If
called with one parameter, the specified number of characters is considered. For example, with no parameters, "Test.java" and "Touring.java" would be placed
together because both start with a T. With 2 as the parameter value, they would be separated ("Te" vs. "To").

If called with 2 parameters, the first parameter is the number of windows to place in one tab, as always (0 meaning to throw them all together).